+++ server/src/test/java/org/opensearch/index/search/stats/SearchStatsTests.java
@@ -32,11 +32,20 @@
 
 package org.opensearch.index.search.stats;
 
+import org.opensearch.action.search.SearchPhase;
+import org.opensearch.action.search.SearchPhaseContext;
+import org.opensearch.action.search.SearchPhaseName;
+import org.opensearch.action.search.SearchRequestStats;
 import org.opensearch.index.search.stats.SearchStats.Stats;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+import static org.hamcrest.Matchers.greaterThanOrEqualTo;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class SearchStatsTests extends OpenSearchTestCase {
 
@@ -63,6 +72,37 @@ public class SearchStatsTests extends OpenSearchTestCase {
         // adding again would then return wrong search stats (would return 4! instead of 3)
         searchStats1.add(searchStats2);
         assertStats(groupStats1.get("group1"), 3);
+
+        long paramValue = randomIntBetween(2, 50);
+
+        // Testing for request stats
+        SearchRequestStats testRequestStats = new SearchRequestStats();
+        SearchPhaseContext ctx = mock(SearchPhaseContext.class);
+        for (SearchPhaseName searchPhaseName : SearchPhaseName.values()) {
+            SearchPhase mockSearchPhase = mock(SearchPhase.class);
+            when(ctx.getCurrentPhase()).thenReturn(mockSearchPhase);
+            when(mockSearchPhase.getStartTimeInNanos()).thenReturn(System.nanoTime() - TimeUnit.SECONDS.toNanos(paramValue));
+            when(mockSearchPhase.getSearchPhaseName()).thenReturn(searchPhaseName);
+            for (int iterator = 0; iterator < paramValue; iterator++) {
+                testRequestStats.onPhaseStart(ctx);
+                testRequestStats.onPhaseEnd(ctx);
+            }
+        }
+        searchStats1.setSearchRequestStats(testRequestStats);
+        for (SearchPhaseName searchPhaseName : SearchPhaseName.values()) {
+            assertEquals(
+                0,
+                searchStats1.getTotal().getRequestStatsLongHolder().getRequestStatsHolder().get(searchPhaseName.getName()).current
+            );
+            assertEquals(
+                paramValue,
+                searchStats1.getTotal().getRequestStatsLongHolder().getRequestStatsHolder().get(searchPhaseName.getName()).total
+            );
+            assertThat(
+                searchStats1.getTotal().getRequestStatsLongHolder().getRequestStatsHolder().get(searchPhaseName.getName()).timeInMillis,
+                greaterThanOrEqualTo(paramValue)
+            );
+        }
     }
 
     private static void assertStats(Stats stats, long equalTo) {
@@ -87,5 +127,4 @@ public class SearchStatsTests extends OpenSearchTestCase {
         // avg_concurrency is not summed up across stats
         assertEquals(1, stats.getConcurrentAvgSliceCount(), 0);
     }
-
 }